pyproject.toml
[tool.poetry]
name = "listen_bravo_test"
version = "0.1.0"
description = "A FastAPI project for processing CSV files and classifying companies"
authors = ["Vijay Singh Kushwaha"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.68.0"
uvicorn = "^0.15.0"
sqlalchemy = "^1.4.23"
python-multipart = "^0.0.5"
langchain = "^0.0.150"
openai = "^0.27.0"
python-dotenv = "^0.19.0"
psycopg2-binary = "^2.9.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"